// src/app/cus9/model/procesoActu.ts
export class ProcesoActu {
  prj_id: number;
  prj_litigacion: string;
  prj_fechanotifica: string;
  prj_numerobizagi: string;
  prj_numerojuzgado: string;
  tij_id: number;
  dei_id: number;
  dpt_idjuzgado: number;
  mpi_idjuzgado: number;
  prj_numeroradica: string;
  prj_23digitos: number;
  tip_id: number;
  prj_clasificacionps: string;
  dem_per_Nombre: string;
  dem_af_numerodocumento: number;
  prj_demandado: number;
  persApo_per_Nombre: string;
  persApo_af_numerodocumento: number;
  per_NumeTarjProfesional: string;
  clt_id: number;
  tii_id: number;
  prj_relacioninforma: string;
  reg_idjuzgado: number;
  prj_anoradicacion: number;
  tiu_id: number;
  tia_id: number;
  prj_hechosproceso: string;
  ead_id: number;
  prj_numresolucion: string;
  cap_id: number;
  mop_id: number;
  prp_id: number;
  clp_id: number;
  prj_fechaadminisiondemanda: string;
  prj_cuantiaestimada: number;
  persAfi_per_Nombre: string;
  persAfi_af_numerodocumento: number;
  persBen_per_Nombre: string;
  perBen_af_numerodocumento: number;
  prj_cantidaddemandante: number;
  prj_entidadpublica: string;
  persApoContra_per_Nombre: string;
  persApoContra_af_numerodocumento: number;
}